+++ tests/Community/Training/Easy/SimpleLoadBalancing/CGTest.php
@@ -1,0 +1,75 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CyrilVerloop\Codingame\Tests\Community\Training\Easy\SimpleLoadBalancing;
+
+use CyrilVerloop\Codingame\Tests\PuzzleTest;
+use CyrilVerloop\Codingame\Community\Training\Easy\SimpleLoadBalancing\SimpleLoadBalancing;
+
+/**
+ * Tests for the "Simple load balancing" puzzle.
+ *
+ * @covers \CyrilVerloop\Codingame\Community\Training\Easy\SimpleLoadBalancing\SimpleLoadBalancing
+ * @group simpleLoadBalancing
+ * @medium
+ */
+final class CGTest extends PuzzleTest
+{
+    public function setUp(): void
+    {
+        $this->puzzle = new SimpleLoadBalancing();
+    }
+
+    /**
+     * Test that the code can be executed for "Test 1".
+     *
+     * @group simpleLoadBalancing_test1
+     */
+    public function testCanExecuteTest1(): void
+    {
+        $this->expectExecuteOutputAnswer(
+            __DIR__ . '/input/01 - test 1.txt',
+            3 . PHP_EOL
+        );
+    }
+
+    /**
+     * Test that the code can be executed for "Test 2".
+     *
+     * @group simpleLoadBalancing_test2
+     */
+    public function testCanExecuteTest2(): void
+    {
+        $this->expectExecuteOutputAnswer(
+            __DIR__ . '/input/02 - test 2.txt',
+            8 . PHP_EOL
+        );
+    }
+
+    /**
+     * Test that the code can be executed for "Test 3".
+     *
+     * @group simpleLoadBalancing_test3
+     */
+    public function testCanExecuteTest3(): void
+    {
+        $this->expectExecuteOutputAnswer(
+            __DIR__ . '/input/03 - test 3.txt',
+            94125 . PHP_EOL
+        );
+    }
+
+    /**
+     * Test that the code can be executed for "Test 4".
+     *
+     * @group simpleLoadBalancing_test4
+     */
+    public function testCanExecuteTest4(): void
+    {
+        $this->expectExecuteOutputAnswer(
+            __DIR__ . '/input/04 - test 4.txt',
+            1 . PHP_EOL
+        );
+    }
+}